requirements.txt
# === additional non-xtts tools ===
python-docx
googletrans==3.1.0a0
transformers
# =================================

torch~=2.4.0
torchaudio~=2.4.0
librosa~=0.10.2.post1
spacy~=3.7.5
tokenizers~=0.19.1
cutlet
opencc
num2words~=0.5.13
transformers~=4.43.3
einops~=0.8.0

numpy~=1.26.4
scipy~=1.13.1
soundfile~=0.12.1
fsspec~=2024.6.1

trainer~=0.0.36
